Add authorizations through public and admin sections

+++ app/controllers/admin/companies_controller.rb
@@ -1,16 +1,18 @@
 class Admin::CompaniesController < ApplicationController
-  before_action :set_company, only: [:destroy]
+  before_action :set_company, only: [:edit, :destroy]
 
   def index
-    @companies = Company.paginate(:page => params[:page], :per_page => 30)
+    @companies = policy_scope(Company).paginate(:page => params[:page], :per_page => 30)
   end
 
   def new
     @company = Company.new
+    authorize @company
   end
 
   def create
     @company = Company.new(company_params)
+    authorize @company
     if @company.save
       redirect_to admin_companies_path
     else
@@ -18,7 +20,18 @@ class Admin::CompaniesController < ApplicationController
     end
   end
 
+  def edit
+    authorize @company
+  end
+
+  def update
+    authorize @company
+    @company.update(company_params)
+    redirect_to admin_companies_path
+  end
+
   def destroy
+    authorize @company
     if @company.destroy
       flash[:notice] = 'L\'entreprise a bien été supprimée'
     else
